requirements.txt
Pillow==9.1.1
progressbar~=2.5